Move get_partition() to enc dir (#5091)
These functions are only needed by the encoder.
