+++ av2/common/av2_common_int.h
@@ -3209,14 +3209,6 @@ static INLINE int frame_is_intra_only(const AV2_COMMON *const cm) {
          cm->current_frame.frame_type == INTRA_ONLY_FRAME;
 }
 
-// Check whether this is chroma component of an intra region in inter frame
-static INLINE int is_inter_sdp_chroma(const AV2_COMMON *const cm,
-                                      REGION_TYPE cur_region_type,
-                                      TREE_TYPE cur_tree_type) {
-  return !frame_is_intra_only(cm) && cur_region_type == INTRA_REGION &&
-         cur_tree_type == CHROMA_PART;
-}
-
 static INLINE int frame_is_sframe(const AV2_COMMON *cm) {
   return cm->current_frame.frame_type == S_FRAME;
 }
@@ -5181,26 +5173,6 @@ static INLINE TX_SIZE get_tx_partition_one_size(TX_PARTITION_TYPE partition,
 }
 
 /*
-Gets the type to signal for the 4 way split tree in the tx partition
-type signaling.
-*/
-static INLINE int get_split4_partition(TX_PARTITION_TYPE partition) {
-  switch (partition) {
-    case TX_PARTITION_NONE:
-    case TX_PARTITION_SPLIT:
-    case TX_PARTITION_VERT:
-    case TX_PARTITION_HORZ:
-    case TX_PARTITION_VERT4:
-    case TX_PARTITION_HORZ4:
-    case TX_PARTITION_HORZ5:
-    case TX_PARTITION_VERT5: return partition;
-    default: assert(0);
-  }
-  assert(0);
-  return 0;
-}
-
-/*
 Gets tx type group table if TX partition supports both vertical and horizontal
 partitions.
 */
@@ -5308,80 +5280,6 @@ static INLINE int use_tx_partition(TX_PARTITION_TYPE partition,
   }
   assert(0);
   return 0;
-}
-
-// Compute the next partition in the direction of the sb_type stored in the mi
-// array, starting with bsize.
-static INLINE PARTITION_TYPE get_partition(const AV2_COMMON *const cm,
-                                           const int plane_type, int mi_row,
-                                           int mi_col, BLOCK_SIZE bsize) {
-  const CommonModeInfoParams *const mi_params = &cm->mi_params;
-  if (mi_row >= mi_params->mi_rows || mi_col >= mi_params->mi_cols)
-    return PARTITION_INVALID;
-
-  const int offset = mi_row * mi_params->mi_stride + mi_col;
-  MB_MODE_INFO **mi = mi_params->mi_grid_base + offset;
-  const BLOCK_SIZE subsize = mi[0]->sb_type[plane_type];
-
-  assert(bsize < BLOCK_SIZES_ALL);
-
-  if (subsize == bsize) return PARTITION_NONE;
-
-  const int bhigh = mi_size_high[bsize];
-  const int bwide = mi_size_wide[bsize];
-  const int sshigh = mi_size_high[subsize];
-  const int sswide = mi_size_wide[subsize];
-
-  if (bsize > BLOCK_8X8 && mi_row + bwide / 2 < mi_params->mi_rows &&
-      mi_col + bhigh / 2 < mi_params->mi_cols) {
-    // In this case, the block might be using an extended partition
-    // type.
-    const MB_MODE_INFO *const mbmi_right = mi[bwide / 2];
-    const MB_MODE_INFO *const mbmi_below = mi[bhigh / 2 * mi_params->mi_stride];
-
-    if (sswide == bwide) {
-      // Smaller height but same width. Is PARTITION_HORZ_4, PARTITION_HORZ or
-      // PARTITION_HORZ_B. To distinguish the latter two, check if the lower
-      // half was split.
-      if (sshigh * 4 == bhigh) {
-        return PARTITION_HORZ_4A;
-      }
-      if (mbmi_below->sb_type[plane_type] == subsize) return PARTITION_HORZ;
-    } else if (sshigh == bhigh) {
-      // Smaller width but same height. Is PARTITION_VERT_4, PARTITION_VERT or
-      // PARTITION_VERT_B. To distinguish the latter two, check if the right
-      // half was split.
-      if (sswide * 4 == bwide) {
-        return PARTITION_VERT_4A;
-      }
-      if (mbmi_right->sb_type[plane_type] == subsize) return PARTITION_VERT;
-
-    } else {
-      // Smaller width and smaller height. Might be PARTITION_SPLIT or could be
-      // PARTITION_HORZ_A or PARTITION_VERT_A. If subsize isn't halved in both
-      // dimensions, we immediately know this is a split (which will recurse to
-      // get to subsize). Otherwise look down and to the right. With
-      // PARTITION_VERT_A, the right block will have height bhigh; with
-      // PARTITION_HORZ_A, the lower block with have width bwide. Otherwise
-      // it's PARTITION_SPLIT.
-      if (sswide * 2 != bwide || sshigh * 2 != bhigh) {
-        if (mi_size_wide[mbmi_below->sb_type[plane_type]] < bwide &&
-            mi_size_high[mbmi_right->sb_type[plane_type]] < bhigh)
-          return PARTITION_SPLIT;
-      }
-      return PARTITION_SPLIT;
-    }
-  }
-  const int vert_split = sswide < bwide;
-  const int horz_split = sshigh < bhigh;
-  const int split_idx = (vert_split << 1) | horz_split;
-  assert(split_idx != 0);
-
-  static const PARTITION_TYPE base_partitions[4] = {
-    PARTITION_INVALID, PARTITION_HORZ, PARTITION_VERT, PARTITION_SPLIT
-  };
-
-  return base_partitions[split_idx];
 }
 
 static AVM_INLINE void av2_set_frame_sb_size(AV2_COMMON *cm,
